Rename script and move helper functions to utils.py
